requirements.txt
fastapi
uvicorn[standard]
pydantic
google-genai
pytest
redis
pypdf
python-docx
python-multipart
